requirements.txt
python-dotenv
openai==0.28
pyyaml
scikit-learn
numpy
pandas
openpyxl
nest_asyncio